+++ rust/src/fdbserver/grv_master.rs
@@ -1,1 +1,2 @@
+#[allow(dead_code)]
 pub fn foo() {}
